refactor Labels category
