+++ pages/movies/movies.module.scss
@@ -6,22 +6,6 @@
   @apply flex-wrap flex justify-center mb-1;
 }
 
-.labelsList {
-  @apply space-y-2 space-x-2 flex-wrap justify-center;
-
-  :first-child {
-    @apply mt-2 ml-2;
-  }
-
-  *:focus {
-    @apply outline-none;
-
-    div {
-      @apply text-red border-red;
-    }
-  }
-}
-
 .loadMore {
   @apply p-3 text-lg font-semibold bg-red rounded;
 }
